requirements.txt
lyricsgenius==3.7.2
python-dotenv==1.0.0

